code/dwh_extractor/requirements.txt
hdbcli
pandas
google-cloud-storage
google-cloud-secret-manager
google-cloud-bigquery
pyarrow
pyodbc
oracledb
sqlserverport
fsspec
gcsfs
openpyxl
